% Generated by roxygen2 (4.0.1): do not edit by hand
\name{becca_plot2.mapvizier}
\alias{becca_plot2.mapvizier}
\alias{beccaplot2}
\title{Wrapper to becca_plot for \code{\link{mapvizier}} objects}
\usage{
\method{becca_plot2}{mapvizier}(.data, ..., subsetting = filter(Year2 >= 2010,
  MeasurementScale \%in\% c("Reading", "Mathematics")))
}
\description{
Wrapper to becca_plot for \code{\link{mapvizier}} objects
}

