% Generated by roxygen2 (4.0.1): do not edit by hand
\name{becca_plot2}
\alias{becca_plot2}
\title{becca_plot2}
\usage{
becca_plot2(.data, ...)
}
\arguments{
\item{.data}{an object to plot}

\item{...}{other arguments passed to \code{\link{becca_plot}}}

\item{subsetting}{(required for \code{mapvizier obejcts}) a call to \code{dplyr}'s \code{filter} function, leavingout
the \code{.data} argument (i.e. only containing subsetting arguments)}
}
\value{
a \code{link[ggplot2]{ggplot2}} object
}
\description{
\code{becca_plot2()} is a method that serves as wrapper for \code{\link{becca_plot}}.
}
\details{
\code{becca_plot2} serves as wrapper that dispatches to \code{\link{becca_plot}} depending
on the class of the object passed to the \code{.data} parameter.

If the object is a \code{\link{mapvizier}} object, then \code{becca_plots} requirement that the user
identify a litany of columns in the data is unnecessary (since we know which
columns are which in a \code{mapvizier object}). The \code{subsetting} argument must be used
when a \code{mapvizier} object is passed, since the data will need to be
properly subsetted. The user should pass this argument a call to \code{\link[dplyr]{filter(...)}}, where
that function's \code{.data} argument is unnecessary. All other arguments (e.g. \code{facets})
are passed to \code{becca_plot} via the \code{...} argument.

Ff the object is a not a \code{mapvizier} object (like a \code{data.frame}) then as with \code{becca_plot}, the
user needs to identify school, cohort, year, grade-level-season, and percentile columns.
}
\seealso{
\code{\link{becca_plot}}, \code{link[dplyr]{filter}}
}

