% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/best_avail_sgp_df.R
\name{best_avail_sgp_df}
\alias{best_avail_sgp_df}
\title{Build a composite data frame of 'best available' SGP}
\usage{
best_avail_sgp_df(df, prefer_period = "Spring to Spring",
  backup_period_1 = "Fall to Spring", backup_period_2 = "Winter to Spring")
}
\arguments{
\item{df}{a data frame of SGPs.  TODO: define structure}

\item{prefer_period}{Name of the top growth term to use.  Default is 'Spring to Spring'.}

\item{backup_period_1}{First backup.  Default is 'Fall to Spring'.}

\item{backup_period_2}{Second backup.  Default is 'Winter to Spring'.}
}
\value{
a data frame of SGPs
}
\description{
\code{best_avail_sgp_df} returns a data frame with one sgp per student, using the growth period
preferences provided.
}
\details{
Default growth period is Spring to Spring, default backup is Fall to Spring.  ie, pull a full
year's growth for most kids; fall back to fall-to-spring for new kids (ie K)
}

