% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calc_growth_targets}
\alias{calc_growth_targets}
\title{for each student.  This is experimental.}
\usage{
calc_growth_targets(.data, typical.growth = TRUE, college.ready = TRUE,
  season1 = "Spring", season2 = "spring")
}
\arguments{
\item{.data}{the data, which is either a \code{mapvizier} object or a \code{data.frame}}

\item{typical.growth}{if \code{TRUE} (the default), thene the function calculates and returns
typcial growth target}

\item{college.ready}{if \code{TRUE} (the default), thene the function calculates and returns
college ready growth target}

\item{season1}{first season to use for growth calculations}

\item{seaosn2}{second season to use for growth calculations}
}
\value{
a data.frame with growth columns or something
}
\description{
\code{cakc_growht_targerts} calculates typical and college ready growth targets.
}
\details{
Growth target calculator
}
\examples{
calc_growth_targets()
}

