% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fall_spring_sort_me}
\alias{fall_spring_sort_me}
\title{Fall-Spring Sort Me}
\usage{
fall_spring_sort_me(grade_level)
}
\arguments{
\item{grade_level}{a grade level between -1 and 12}
}
\value{
a string
}
\description{
\code{fall_spring_me} tranforms grade levels into labels that will sort properly;
eg, fall / winter/ spring
}

