% Generated by roxygen2 (4.0.1): do not edit by hand
\name{galloping_elephants}
\alias{galloping_elephants}
\title{RIT distribution change (affectionately titled 'Galloping Elephants')}
\usage{
galloping_elephants(df, first_and_spring_only = TRUE, entry_grades = c(-0.7,
  4.3), title_text = FALSE)
}
\arguments{
\item{df}{long data frame in TEAM canonical style}

\item{first_and_spring_only}{should fall/winter scores be shown, or only endpoint scores?}

\item{entry_grades}{the entry grades that are considered 'first' for a cohort (default is KF and 5F)}

\item{title_text}{desired plot title}
}
\value{
returns a ggplot object
}
\description{
\code{one_year_sgp} returns ggplot density distributions that show change in RIT over time
}

