% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/galloping_elephants_plot.R
\name{galloping_elephants_plot.default}
\alias{galloping_elephants_plot.default}
\title{Wrapper to galloping_elephants for non \code{\link{mapvizier}} objects}
\usage{
\method{galloping_elephants_plot}{default}(.data,
  first_and_spring_only = TRUE, entry_grades = c(-0.8, 4.2),
  title_text = FALSE, ho_cutoff = 13, bw_adjust = 1,
  school_name_column = "sch_abbrev",
  academic_year_column = "map_year_academic",
  grade_level_season = "grade_level_season",
  measurement_scale_column = "measurementscale",
  rit_score_column = "testritscore",
  test_percentile_column = "percentile_2011_norms",
  current_grade_column = "cur_grade_level")
}
\description{
Wrapper to galloping_elephants for non \code{\link{mapvizier}} objects
}

