% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/galloping_elephants_plot.R
\name{galloping_elephants_plot.mapvizier}
\alias{galloping_elephants_plot.mapvizier}
\title{Wrapper to galloping_elephants for \code{\link{mapvizier}} objects}
\usage{
\method{galloping_elephants_plot}{mapvizier}(.data, ...,
  subsetting = filter(CohortYear == 2023, MeasurementScale == "Reading",
  Season != "Winter"))
}
\description{
Wrapper to galloping_elephants for \code{\link{mapvizier}} objects
}

