% Generated by roxygen2 (4.0.1): do not edit by hand
\name{glimpse}
\alias{glimpse}
\title{glimpse method for \code{mapvizier} class}
\usage{
glimpse(x)
}
\arguments{
\item{x}{a \code{mapvizier} object}
}
\value{
a list of glimpses to the console
}
\description{
prints to console glimpse of a the first lines of a \code{mapvizier} object's
 components.
}
\details{
Essentialy extendls Hadley Wickham's \code{glimpse} function from
the \code{dplyr} package to work with \code{mapvizier} objects.  \code{mapvizier}
objects are lists of two data frames (\code{mapData} and \code{seasonMatched}). This
method works over both data frames.
}
\examples{
data(nweamap)
x<-mapvizier(nweamap)
glimpse(x)
}

