% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grade_level_season.R
\name{grade_level_season}
\alias{grade_level_season}
\title{Grade Level Season}
\usage{
grade_level_season(fws_string)
}
\arguments{
\item{fws_string}{Fall, Winter or Spring}
}
\description{
\code{grade_level_season} takes a map_season and returns a decimal representation of
what approximate portion of the year it represents
}

