% Generated by roxygen2 (4.0.1): do not edit by hand
\name{haid_plot}
\alias{haid_plot}
\title{Chris Haid's Waterfall-Rainbow-Arrow Chart}
\usage{
haid_plot(df, student_col_name, start_col_name, end_col_name,
  start_percentile_col_name, end_percentile_col_name, keep_up_col_name,
  col_ready_col_name, start_season_abbrev, end_season_abbrev,
  growth_status_col_name, name_status_col_name, y_sort_column, p_title = "",
  p_arrow_colors, p_arrow_tiers, p_name_colors, p_name_color_tiers,
  p_quartile_colors = c("#f3716b", "#79ac41", "#1ebdc2", "#a57eb8"),
  p_name_size = 3, p_alpha = 1)
}
\arguments{
\item{df}{data.frame with columns for  RIT scores (start and/or end season), predicted typical and college ready
growth, growth status, etc.}

\item{student_col_name}{string identifying colummn in \code{df} of student names.}

\item{start_col_name}{string identifying colummn in \code{df} of first term's RIT score.}

\item{end_col_name}{string identifying colummn in \code{df} of second term's RIT score.}

\item{start_percentile_col_name}{string identifying colummn in \code{df} with first terms NPR.}

\item{end_percentile_col_name}{string identifying colummn in \code{df} with second terms NPR.}

\item{keep_up_col_name}{string identifying colummn in \code{df} typical (or project) term 1 to term 2 growth.}

\item{col_ready_col_name}{string identifying colummn in \code{df} college ready term 1 to term 2 growth}

\item{start_season_abbrev}{string indicating term 1's name (e.g. Fall).}

\item{end_season_abbrev}{string indicating term 2's name (e.g. Spring).}

\item{growth_status_col_name}{string identifying column in \code{df} indicating students growth status.}

\item{name_status_col_name}{??? Thoughts Andrew???}

\item{y_sort_column}{string identifying column in \code{df} used to sort students vertically.}

\item{p_title}{= '' string used for title.}

\item{p_arrow_colors}{vector of colors passed to ggplot.}

\item{p_arrow_tiers}{vector of colors passed to ggplot.}

\item{p_name_colors}{vector of colors passed to ggplot.}

\item{p_name_color_tiers}{vector of colors passed to ggplot.}

\item{p_quartile_colors}{vector of colors passed to ggplot.}

\item{p_name_size}{sets point size of student names.}

\item{p_alpha}{sets level fo transparency for goals.}
}
\value{
prints a ggplot object
}
\description{
\code{haid_plot} returns a ggplot object showing per student MAP performance
}
\details{
This function builds and prints aa graphich that plots MAP performance over one
or two seasons. RIT scores are color coded by percentile. If a second season is passed to the function (in \code{end_col_name}),
then arrows between estimated RIT scores for the two seasons are added to plot and are color coded
based on the amount of growth attained (postive, typical, college ready, and negative)
}
\examples{
haid_plot(...)
}

