% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/height_weight_npr.R
\name{height_weight_npr}
\alias{height_weight_npr}
\title{RIT "Height Weight" Growth, with NPR bands}
\usage{
height_weight_npr(desired_subj, norms_dense, color_list = rainbow_colors(),
  ribbon_alpha = 0.35, annotation_style = "points", line_style = "none")
}
\arguments{
\item{desired_subj}{}
}
\value{
a ggplot object
}
\description{
\code{height_weight_npr} returns a ggplot object meant to be used as a template
}

