% Generated by roxygen2 (4.0.1): do not edit by hand
\name{join_map_files}
\alias{join_map_files}
\title{join_map_files}
\usage{
join_map_files(assessment_results, students_by_school)
}
\arguments{
\item{assessment_results}{}

\item{students_by_school}{}
}
\description{
\code{join_map_files} combines AssessmentResults and StudentsBySchool
}

