% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{mapvisuals}
\alias{mapvisuals}
\alias{mapvisuals-package}
\title{KIPP MAP Visuals}
\description{
Functions that support analysis of NWEA MAP assessment data
}
\details{
The mapvisuals package.
}
\author{
Andrew Martin

Chris Haid
}

