% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mapvizier}
\alias{mapvizier}
\alias{mapvizier.default}
\title{Create a mapvizier object}
\usage{
mapvizier(x)

\method{mapvizier}{default}(x)
}
\arguments{
\item{x}{a data frame from a long CDF that has student data
already joined to it}
}
\value{
a \code{mapvizier} object
}
\description{
\code{mapvizier()} returns an object with class \code{mapvizier}
}
\details{
This method takes a long form CDF, addess columns for growth calculations and
subsetting and joins assessemnts by student ID and measurement scale for seasons
in appropriate school years.
}
\examples{
require(mapvizier)

data(nweamap)
x<-mapvizier(nweamap)
str(x)
}

