% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/network_typ_growth_distro.R
\name{network_typ_growth_distro}
\alias{network_typ_growth_distro}
\title{Network Typical Growth Distribution}
\usage{
network_typ_growth_distro(nat_results_df, measurementscale, academic_year,
  grade_level, start_season, end_season, comparison_name,
  comparison_pct_typ_growth, replace_nat_results_match = FALSE,
  de_kippify_names = TRUE, de_schoolify_names = TRUE)
}
\arguments{
\item{nat_results_df}{KIPP results nationwide}
}
\value{
a ggplot chart
}
\description{
\code{network_typ_growth_distro} shows your school relative to other KIPP schools
}

