% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{nweamap}
\alias{nweamap}
\title{NWEA MAP Colllection of Comprehensive Data File}
\format{a \code{data.frame}with 22607 obs. of  72 variables:}
\usage{
nweamap
}
\description{
Mulitple years' of multiples schools' worth of NWEA data for a single
district
}
\author{
Christopher J. Haid
}
\keyword{datasets}

