% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.mapvizierSummary}
\alias{plot.mapvizierSummary}
\title{plot method for \code{mapvizierSummary} class}
\usage{
\method{plot}{mapvizierSummary}(x, growthseason = "Fall - Spring",
  subjects = c("Reading", "Mathematics", "General Science"),
  grades = c(2:8), years = NULL, metric = "Typical")
}
\arguments{
\item{x}{a \code{mapvizierSummary} object}

\item{growthseason}{a character string denoting test term name (i.e. growth seaons):
\itemize{
\item "Spring - Spring"
\item "Fall - Spring"
\item "Fall - Winter"
\item "Winter - Spring"
\item "Fall - Fall"
}}

\item{subjects}{character string to subset subjects/measurement scales by:
\itemize{
\item "Mathematics"
\item "Reading"
\item "General Science"
\item "Language Usage"
}}

\item{grades}{integer vector to subset grades by (NB: Kindergarten should be denoted 0)}

\item{years}{integer vector with years of second test year in term names (i.e.,
to plot results from SY2010-2011, SY 2011-2012, and SY 2012-2013, you should pass the
vector \code(c(2011, 2012, 2013)) or \code(c(2011:2013))).}

\item{metric}{character value indicating which MAP measurement to plot. Choices are:
\itemize{
\item "Typical": % Meets/Exceeds\nTypical Growth,
\item "CR": % Meets/Exceeds\nCollege Ready Growth,
\item "50th" = % Greater than\n50th Percentile, Season 2,
\item "75th" = % Greater than\75th Percentile, Season 2.
}}
}
\value{
a \code{gglot2} object
}
\description{
Plots a \code{ggplot2} object from a \code{mapvizierSummary} object
}
\details{
Creates and prints a \code{ggplot2} object from
a \code{mapvizierSummary} object.  User can choose subjects, growth season,
grades, and years to subest data prior to plotting
}
\examples{
data(nweamap)
x<-mapvizier(nweamap)
x.sum<-summary(x)
plot(x.summ)
}

