% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.mapvizier.R
\name{print.mapvizier}
\alias{print.mapvizier}
\title{print method for \code{mapvizier} class}
\usage{
\method{print}{mapvizier}(x, ...)
}
\arguments{
\item{x}{a \code{mapvizier} object}
}
\value{
some details about the object to the console.
}
\description{
prints to console
}
\details{
Prints a summary fo the a \code{mapvizier} object.
}
\examples{
data(nweamap)
x<-mapvizier(nweamap)
x
}

