% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sgp_small_multiples.R
\name{sgp_small_multiples}
\alias{sgp_small_multiples}
\title{SGP small multiples}
\usage{
sgp_small_multiples(df, stu_per_row = 12)
}
\arguments{
\item{df}{long data frame in TEAM canonical style}

\item{stu_per_row}{how many kids wide?}
}
\value{
returns a ggplot object
}
\description{
\code{sgp_small_multiples} returns a ggplot facet_grid with a box for every kid, showing
their SGP and rit change
}

