% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/strands_plot.R
\name{strands_plot}
\alias{strands_plot}
\title{Plot MAP Goal Strand Results}
\usage{
strands_plot(.data, ...)
}
\arguments{
\item{...}{arguments passed to \code{\link[dplyr]{filter}} use to subset \code{x}.}

\item{x}{a \code{mapvizier} object or data.frame with the following columns:
\itemize{
 \item  StudentID,
 \item StudentFirstname,
 \item StudentLastname,
 \item SchoolInitials,
 \item Grade,
 \item MeasurementScale,
 \item TestRITScore,
 \item TestPercentile,
 \item TestQuartile,
 \item Classroom,
 \item Year2,
 \item Season
 \item columns with of Goal Strand RIT score of interest named Goal1RitScore, Goal2RitScore, etc.
 \item columns with of Goal Strand Names of interest named Goal1Name, Goal2Name, etc.
}}
}
\value{
a ggplot2 object
}
\description{
Plots a group of students' goal strand RIT scores, with each
student ranked by overall RIT score.  Colors of points indicating goal strand RIT
scores indicate deviation from overall RIT score; colors of overall RIT mark (|)
indicate national percentile rank of the overall RIT score
}
\details{
Creates and prints a ggplot2 object showing both overall and goal strand RIT
scores for each student in a subets.  Subsetting students is very imporant for
making this visualization useful.  Original data from a \code{mapvizier} object
can be subsetted by by passing unquoted field names to the \code{...} argument, a
al \code{\link[dplyr]{filter}}, to which these arguments are passed. See the example
below.
}
\examples{
data(nweamap)
x<-mapvizier(nweamap)
strands_plot(x,
             Grade==8,
             Classroom=="UChicago",
             MeasurementScale=="Mathematics",
             Year2==2014,
             Season=="Spring")
}

