% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/term_fallwinterspring.R
\name{term_fallwinterspring}
\alias{term_fallwinterspring}
\title{term_fallwinterspring}
\usage{
term_fallwinterspring(term_names)
}
\arguments{
\item{term_names}{a vector of term names}
}
\description{
\code{term_fallwinterspring} takes a MAP CDF and returns munges the termname field to
return season and year seperately.
}

