% Generated by roxygen2 (4.0.1): do not edit by hand
\name{term_map_year_academic}
\alias{term_map_year_academic}
\title{term_map_year_academic}
\usage{
term_map_year_academic(term_names)
}
\arguments{
\item{term_names}{a vector of term names}
}
\description{
\code{term_map_year_academic} takes a MAP CDF and returns munges the termname field to
return season and year seperately.
}

