% Generated by roxygen2 (4.0.1): do not edit by hand
\name{term_split}
\alias{term_split}
\title{term_split}
\usage{
term_split(regular_cdf)
}
\arguments{
\item{regular_cdf}{a NWEA MAP cdf loaded in as a data frame}
}
\description{
\code{term_splot} takes a MAP CDF and returns munges the termname field to
return season and year seperately.
}

