% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/term_split2.R
\name{term_split2}
\alias{term_split2}
\title{Split vector like NWEA CDF TermName column into usefule components}
\usage{
term_split2(term_name, grade = NULL)
}
\arguments{
\item{term_name}{character vector of term names from NWEA CDF.}

\item{grade}{(optional, default = \code{NULL}) integer vector of grade levels}
}
\value{
a data frame  of with \code{length(x)} rows and 4 (5 if grade vector supplied) columns
}
\description{
\code{term_split2} returns a data frame splitting a TermName
}
\details{
This function returns a data framre with \code{length(x)} rows that separates a TermName character
vector into the following components (here we use "Spring 2012-2013" as en example
element of the vector:
\itemize{
\item  Season ("Spring"),
\item  Year1 (2012),
\item  Year2 (2013),
\item  SY (i.e School Year, "2013-2014").
}
In addition, if a \code{length(x)} vector of grade levels is passed to the optional second argument
then an additional column will be added to the data frame:
\itemize{
\item  CohortYear (HS graduation year based on grade and year2, so an 7th grader in spring 2013 would be
in the class of 2018).
}
}
\examples{
x<-c("Spring 2012-2013",
"Fall 2013-2014",
     "Winter 2012-2013",
     "Fall 2012-2013" ,
     "Spring 2013-2014")

grades<-c(3:7)
term_split2(term_name=x,
            grade=grades )
}

