% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/becca_plot.R
\name{becca_plot}
\alias{becca_plot}
\title{Becca Vichniac's Quartile (Floating Bar) Chart}
\usage{
becca_plot(.data, school_name_column = "sch_abbr",
  cohort_name_column = "cohort",
  grade_level_season_column = "grade_level_season",
  measurement_scale_column = "measurementscale",
  test_percentile_column = "percentile_2011_norms",
  first_and_spring_only = TRUE, entry_grades = c(-0.8, 4.2),
  auto_justify_x = TRUE, justify_widths = FALSE, justify_min = NA,
  justify_max = NA, color_scheme = "KIPP Report Card", facets = FALSE,
  facet_opts = FALSE, title_text = FALSE, small_n_cutoff = 0.001)
}
\arguments{
\item{.data}{the data frame in TEAM canoncical style (long data forms)}

\item{school_name_column}{column in \code{.data} with school names}

\item{cohort_name_column}{column in \code{.data} with cohornt names}

\item{grade_level_season_column}{column in \code{.data} with numeric indicating grade season (e.g., Fall 4th
= 3.3, Winter 4th = 3.7, Spring 4th = 4.0).}

\item{measurement_scale_column}{column in \code{.data} with subject}

\item{test_percentile_column}{= column in \code{.data} with NPR.}

\item{first_and_spring_only}{indicator for showing shoing Fall-to-Spring rather than Spring-to-Spring}

\item{entry_grades}{= c(-0.7, 4.3)  GradeSeasons that should use Fall and Spring (i.e., Kinder =-.3 and 5th = 4.3)}

\item{justify_widths}{width justification indicator}

\item{justify_min}{}

\item{justify_max}{}

\item{color_scheme}{only 'KIPP Report Card'}

\item{facets}{= FALSE}

\item{facet_opts}{= FALSE}

\item{title_text}{= FALSE}

\item{small_n_cutoff}{drop a grade_level_season if less than x% of the max?
(useful when dealing with weird cohort histories)}
}
\value{
prints a ggplot object
}
\description{
\code{becca_plot} returns a ggplot object binned quaritle performonce
}
\details{
This function builds and prints a bar graph with 4 bins per bar show MAP data
binned by quartile (National Percentile Rank).  Bars are centered at 50th percentile
horizonatally
}

