% Generated by roxygen2 (4.0.1): do not edit by hand
\name{becca_plot2}
\alias{becca_plot2}
\title{becca_plot2}
\usage{
becca_plot2(.data, ...)
}
\arguments{
\item{x}{an object to plot}

\item{...}{other arguments passed to \code{\link{becca_plot}}}
}
\value{
a \code{link[ggplot2]{ggplot2}} object
}
\description{
method as wrapper for \code{\link{becca_plot}}
}
\details{
See \code{\link{becca_plot}}
}

