% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dots}
\alias{dots}
\title{An internal utility function to capture dots (\code{...}) arguments.}
\usage{
dots(...)
}
\arguments{
\item{\code{...}}{arguments to be caputred}
}
\value{
a list of captured dots
}
\description{
\code{dots} capturees dots
}

