% Generated by roxygen2 (4.0.1): do not edit by hand
\name{growth_status_scatter}
\alias{growth_status_scatter}
\title{Growth vs Percentile Rank scatterplot}
\usage{
growth_status_scatter(df, show_title = FALSE, school, subject, grade)
}
\arguments{
\item{df}{long data frame in TEAM canonical style}

\item{show_title}{logical; should the plot have a title?}

\item{school}{}

\item{subject}{}

\item{grade}{}
}
\value{
returns a ggplot object
}
\description{
\code{growth_status_scatter} returns a ggplot scatterplot that shows student growth vs performance
}

