% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/join_map_files.R
\name{join_map_files}
\alias{join_map_files}
\title{join_map_files}
\usage{
join_map_files(assessment_results, students_by_school)
}
\arguments{
\item{assessment_results}{}

\item{students_by_school}{}
}
\description{
\code{join_map_files} combines AssessmentResults and StudentsBySchool
}

