% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/one_year_sgp.R
\name{one_year_sgp}
\alias{one_year_sgp}
\title{One Year SGP histogram}
\usage{
one_year_sgp(df, title_text = FALSE)
}
\arguments{
\item{df}{long data frame in TEAM canonical style}

\item{title_text}{desired plot title}
}
\value{
returns a ggplot object
}
\description{
\code{one_year_sgp} returns a ggplot histogram that shows the distribution of SGPs
}

