% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ordinal_me.R
\name{ordinal_me}
\alias{ordinal_me}
\title{Ordinal me}
\usage{
ordinal_me(number)
}
\arguments{
\item{number}{a number}
}
\value{
returns a string
}
\description{
\code{ordinal_me} appends the appropriate ending ('st', 'nd' etc) to make an integer into an ordinal.
}

