% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prep_cdf}
\alias{prep_cdf}
\title{Prep CDF}
\usage{
prep_cdf(regular_cdf)
}
\arguments{
\item{regular_cdf}{a NWEA MAP cdf loaded in as a data frame}
}
\description{
\code{prep_cdf} is a wrapper around \code{term_fallwinterspring} and \code {term_map_year_academic}
}

