% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prep_cdf.R
\name{prep_cdf}
\alias{prep_cdf}
\title{Prep CDF}
\usage{
prep_cdf(regular_cdf)
}
\arguments{
\item{regular_cdf}{a NWEA MAP cdf loaded in as a data frame}
}
\description{
\code{prep_cdf} is a wrapper around the term functions
}

