% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sgp_small_multiples_cleanup.R
\name{sgp_small_multiples_cleanup}
\alias{sgp_small_multiples_cleanup}
\title{SGP small multiples cleanup}
\usage{
sgp_small_multiples_cleanup(plot)
}
\arguments{
\item{plot}{a ggplot created by sgp_small_multiples}
}
\value{
returns a ggplot object
}
\description{
\code{sgp_small_multiples_cleanup} is a helper function that makes the legend show correctly
(size, show as number, etc)
}

