% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/strand_boxes.R
\name{strand_boxes}
\alias{strand_boxes}
\title{Strand boxes}
\usage{
strand_boxes(df)
}
\arguments{
\item{df}{long data frame in TEAM canonical style}
}
\value{
returns a ggplot object
}
\description{
\code{strand_boxes} returns a ggplot boxplot of MAP strand scores
}

